weather_app/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
requests
apscheduler
matplotlib
plotly
python-dotenv
